feat: add PostgreSQL support for CI tests and update test database configuration (#6)

+++ tests/conftest.py
@@ -3,6 +3,7 @@ import tempfile
 from pathlib import Path
 
 import pytest
+import sqlalchemy as sa
 from alembic import command
 from alembic.config import Config
 from api.deps import get_db_session
@@ -11,21 +12,38 @@ from sqlmodel import Session, create_engine
 from starlette.testclient import TestClient
 
 
+def pytest_addoption(parser):
+    """Add custom command line options."""
+    parser.addoption(
+        "--db",
+        action="store",
+        default="sqlite",
+        choices=["sqlite", "postgres"],
+        help="Database to use for tests: sqlite (default) or postgres",
+    )
+
+
 @pytest.fixture
 def data_folder_path():
     return Path(__file__).parent.parent / "data"
 
 
 @pytest.fixture(name="session")
-def session_fixture(monkeypatch) -> Session:
+def session_fixture(request, monkeypatch) -> Session:
     """Create a new database session for a test."""
     # TODO: Use an in-memory SQLite database for faster tests if possible.
     #       https://sqlmodel.tiangolo.com/tutorial/fastapi/tests/#memory-database
 
-    # Create a temporary database file
-    fd, path = tempfile.mkstemp(suffix=".db")
-    os.close(fd)
-    db_url = f"sqlite:///{path}"
+    db_type = request.config.getoption("--db")
+
+    if db_type == "postgres":
+        db_url = "postgresql://develop:develop_secret@localhost:5432/develop"
+        path = None
+    else:
+        # Create a temporary database file for SQLite
+        fd, path = tempfile.mkstemp(suffix=".db")
+        os.close(fd)
+        db_url = f"sqlite:///{path}"
 
     # Use monkeypatch to set DATABASE_URL environment variable
     monkeypatch.setattr("config.settings.DATABASE_URL", db_url)
@@ -42,12 +60,31 @@ def session_fixture(monkeypatch) -> Session:
     engine = create_engine(db_url)
     # TODO: Alternatively, you can create tables directly without migrations for simpler setups.
     # create_db_and_tables(engine)
-    with Session(engine) as session:
-        yield session
+
+    # Ensure that changes made during tests do not persist and affect other tests using a nested transaction
+    # This is needed for PostgreSQL since the SQLite is erased after each test by deleting the temp file
+    connection = engine.connect()
+    transaction = connection.begin()
+    session = Session(bind=connection)
+
+    nested = connection.begin_nested()
+
+    @sa.event.listens_for(session, "after_transaction_end")
+    def end_savepoint(session, transaction):
+        nonlocal nested
+        if not nested.is_active:
+            nested = connection.begin_nested()
+
+    yield session
+
+    session.close()
+    transaction.rollback()
+    connection.close()
 
     # Clean up
     engine.dispose()
-    os.unlink(path)
+    if path:
+        os.unlink(path)
 
 
 @pytest.fixture(name="client_with_db")
